ml_service/pyproject.toml
[project]
name = "ml-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "librosa>=0.11.0",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "soundfile>=0.13.1",
    "uvicorn>=0.37.0",
    "yt-dlp>=2025.10.22",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = [
    "extractors",
    "models",
    "repositories"
]


